Debug display of phase value in control labels

+++ MatrixTransformations/MatrixForm.cs
@@ -173,7 +173,7 @@ public partial class MatrixForm : Form
         if (_shouldEndPhaseAnimation)
         {
             _phase = Phase.Two;
-            phaseValue.Invoke((MethodInvoker)(() => phaseValue.Text = _phase.ToString()));
+            phaseValue.Invoke((MethodInvoker)(() => phaseValue.Text = ((int) _phase).ToString()));
 
             _shouldEndPhaseAnimation = false;
             _shouldPlayPhaseAnimationForward = true;
@@ -211,7 +211,7 @@ public partial class MatrixForm : Form
         if (_shouldEndPhaseAnimation)
         {
             _phase = Phase.Three;
-            phaseValue.Invoke((MethodInvoker)(() => phaseValue.Text = _phase.ToString()));
+            phaseValue.Invoke((MethodInvoker)(() => phaseValue.Text = ((int) _phase).ToString()));
 
             _shouldEndPhaseAnimation = false;
             _shouldPlayPhaseAnimationForward = true;
@@ -249,7 +249,7 @@ public partial class MatrixForm : Form
         if (_shouldEndPhaseAnimation)
         {
             _phase = Phase.Four;
-            phaseValue.Invoke((MethodInvoker)(() => phaseValue.Text = _phase.ToString()));
+            phaseValue.Invoke((MethodInvoker)(() => phaseValue.Text = ((int) _phase).ToString()));
 
             _shouldEndPhaseAnimation = false;
             _shouldPlayPhaseAnimationForward = true;
@@ -274,7 +274,7 @@ public partial class MatrixForm : Form
         if (_phi == DefaultPhi && _theta == DefaultTheta)
         {
             _phase = Phase.One;
-            phaseValue.Invoke((MethodInvoker)(() => phaseValue.Text = _phase.ToString()));
+            phaseValue.Invoke((MethodInvoker)(() => phaseValue.Text = ((int) _phase).ToString()));
             return;
         }
 
@@ -555,7 +555,7 @@ public partial class MatrixForm : Form
         distanceValue.Text = _distance.ToString(CultureInfo.InvariantCulture);
         phiValue.Text = _phi.ToString(CultureInfo.InvariantCulture);
         thetaValue.Text = _theta.ToString(CultureInfo.InvariantCulture);
-        phaseValue.Text = _phase.ToString();
+        phaseValue.Text = ((int) _phase).ToString();
         animationValue.Text = "Play";
     }
 }
